requirements.txt
pytorch
loguru
